// src/ppms/types/propertyDetails.ts
export interface PropertyDetails {
    data: {
        createdAt: Date;
        updatedAt: Date;
        createdBy: string;
        updatedBy: string;
        propertyId: number;
        itemControlNumber: string;
        aacId: string;
        agencyBureau: string;
        submittedDate: Date;
        submittedBy: string;
        notify_poc: boolean;
        propertyRegion: string;
        propertyStatus: PropertyStatus;
        airCraft: Record<string, unknown> | null;
        vehicle: Record<string, unknown> | null;
        weapon: Record<string, unknown> | null;
        vessel: Record<string, unknown> | null;
        computer: Record<string, unknown> | null;
        trailerHome: Record<string, unknown> | null;
        reportingAgencyAddress: PropertyLocation;
        propertyLocation: PropertyLocation;
        propertyPOC: PropertyContact;
        propertyCustodian: PropertyContact;
        uploadItemList: any[];
        propertyNotes: any[];
        make: string;
        model: string;
        propertyType: string;
        contractInventoryCode: string;
        overseasInventoryCode: string;
        agencyLocationCode: string;
        agencyControlNumber: string;
        amountTobeReimbursed: string;
        manufacturer: string;
        manufactureDate: null;
        federalSalesCenter: string;
        excessReleaseDate: null;
        internalScreeningStartDate: null;
        surplusReleaseDate: Date;
        cflScreeningStartDate: null;
        cflReleaseDate: null;
        externalScreeningStartDate: Date;
        availableInSalesDate: null;
        acquisitionDate: null;
        fscCode: string;
        niinCode: string;
        itemName: string;
        specialDescriptionCode: string;
        specialDescriptionText: string;
        quantity: number;
        quantityReported: number;
        unitOfIssue: string;
        originalAcquisitionCost: number;
        totalAcquisitionCost: number;
        fairMarketValue: null;
        supplyConditionCode: string;
        conditionCode: string;
        hazardous: string;
        fscapCode: string;
        demilitarizationCode: string;
        isDeleted: boolean;
        propertyDescription: string;
        isSubmitted: boolean;
        notifyCustodian: boolean;
        valueAddedServices: null;
        isDonation: null;
        isExchangeSale: string;
        reimbursementRequiredFlag: string;
        reimbursementCode: null;
        dropAfterInternalScreening: boolean;
        withDrawnDate: null;
        withDrawnReason: null;
        withDrawnBy: null;
        rejectedDate: null;
        rejectedReason: null;
        rejectedBy: null;
        destroyedDate: null;
        destroyedReason: null;
        destroyedBy: null;
        categoryCode: number;
        salesItemName: null;
        categoryName: string;
        sourceCode: string;
        plantClearanceLineNumber: number;
        plantClearanceReferenceNumber: string;
        plantClearanceCaseNumber: string;
        partNumber: null;
        drmoCode: string;
        propertyCreationSource: string;
        recipientName: null;
        propertyGroup: string;
        donorInfo: null;
        recipientInfo: null;
        actionCode: null;
        salesCenter: null;
        assignedScoEmail: null;
        assignedMktSpclEmail: null;
        nasaItemIndicator: null;
        appraisalInfo: null;
        appraisalAgencyInfo: null;
        giftInfo: null;
        withdrawalComment: null;
        dosApproverName: null;
        dosApprovalDate: null;
        recalledDate: null;
        recalledReason: null;
        recalledBy: null;
        lastInventoryDate: null;
        lastInventoriedBy: null;
        siteStorage: string;
        countryCode: null;
        salesQuantity: null;
        salesUnitOfIssue: null;
        salesOac: null;
        salesTotalOac: null;
        salesPropertyDescription: null;
        count: null;
        oldSrdDate: null;
        isInternal: boolean;
        quantityRequested: number;
        isChangeRequestRequired: null;
        oldSRDValue: null;
        editDocumentsFlag: boolean;
        salesNotes: null;
        categoryCodeCount: null;
        vin: null;
        fgAPOContact: null;
    }
}

export interface PropertyContact {
    createdAt: Date;
    updatedAt: Date;
    createdBy: string;
    updatedBy: string;
    contactId: number;
    firstName: string;
    lastName: string;
    middleName: null;
    email: string;
    ccEmail: null | string;
    phone: number;
    fax: number;
    phoneExtension: null;
    isDeleted: boolean;
}

export interface PropertyLocation {
    createdAt: Date;
    updatedAt: Date;
    createdBy: string;
    updatedBy: string;
    addressId: number;
    line1: string;
    line2: string;
    line3: string;
    city: string;
    stateCode: string;
    zip: string;
    zip2: null;
    overseasZip: null;
    isDeleted: boolean;
    instructions: null;
    latitude: number | null;
    longitude: number | null;
}

export interface PropertyStatus {
    createdAt: Date;
    updatedAt: Date;
    createdBy: string;
    updatedBy: string;
    statusId: number;
    statusName: string;
    statusDescription: string;
}